+++ sq/CHANGELOG.md
@@ -15,12 +15,12 @@
   `ginputDefault` and `goutputDefault` when possible as default
   derivation records and sums of records deriving GHC's `Generic`.
 
-* Added `EncodeDefault` instances for `aeson`'s `Encoding and `Value`,
+* Added `EncodeDefault` instances for `aeson`'s `Encoding`, `Value` and `Key`,
   for `binary`'s `Put`, for `UUID`, for `Scientific`, for `Fixed`, for
   `network-uri`'s `URI`.
 
-* Added `DecodeDefault` instances for `UUID`, for `aeson`'s `Value`, for
-  `Scientific`, for `Fixed`, for `network-uri`'s `URI`.
+* Added `DecodeDefault` instances for `UUID`, for `aeson`'s `Value` and `Key`,
+  for `Scientific`, for `Fixed`, for `network-uri`'s `URI`.
 
 * Export `BindingName` constructor.